#60 - fix buffer overflow in DenoiseSharpen node
pixel offset computation are wrong in second image loop
during the rendering of the plugin.
This causes a negative pixel offest on a unsigned int value, so overflow.

This patch fixes the overflow, but the rendering for some zoom factor
is not correct, rendering some black area than the modified image.
This happen only when the srcWindow size is lower than the rendering
Window.
Note: in Natron master branch, these sizes are always the same,
hidding the problem of the rendering loop computation.

Signed-off-by: Guillaume Roguez <yomgui1@gmail.com>